# backend/openapi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "RunCoach.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/onboarding/turns": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingTurnRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingTurnRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingTurnRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingTurnResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingTurnResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingTurnResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/onboarding/state": {
      "get": {
        "tags": [
          "Onboarding"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/onboarding/answers/revise": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseAnswerRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseAnswerRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseAnswerRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/onboarding/answers": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitStructuredAnswersRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitStructuredAnswersRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitStructuredAnswersRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStateDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OnboardingProgressDto": {
        "required": [
          "completedTopics",
          "totalTopics"
        ],
        "type": "object",
        "properties": {
          "completedTopics": {
            "type": "integer",
            "format": "int32"
          },
          "totalTopics": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "OnboardingStateDto": {
        "required": [
          "completedTopics",
          "currentFitness",
          "currentTopic",
          "injuryHistory",
          "isComplete",
          "outstandingClarifications",
          "preferences",
          "primaryGoal",
          "status",
          "targetEvent",
          "totalTopics",
          "userId",
          "weeklySchedule"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/OnboardingStatus"
          },
          "currentTopic": {
            "$ref": "#/components/schemas/OnboardingTopic"
          },
          "completedTopics": {
            "type": "integer",
            "format": "int32"
          },
          "totalTopics": {
            "type": "integer",
            "format": "int32"
          },
          "isComplete": {
            "type": "boolean"
          },
          "outstandingClarifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OnboardingTopic"
            }
          },
          "primaryGoal": {
            "$ref": "#/components/schemas/PrimaryGoalAnswer"
          },
          "targetEvent": {
            "$ref": "#/components/schemas/TargetEventAnswer"
          },
          "currentFitness": {
            "$ref": "#/components/schemas/CurrentFitnessAnswer"
          },
          "weeklySchedule": {
            "$ref": "#/components/schemas/WeeklyScheduleAnswer"
          },
          "injuryHistory": {
            "$ref": "#/components/schemas/InjuryHistoryAnswer"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesAnswer"
          },
          "currentPlanId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OnboardingStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "OnboardingTopic": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "OnboardingTurnKind": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "OnboardingTurnRequestDto": {
        "required": [
          "idempotencyKey",
          "text"
        ],
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "OnboardingTurnResponseDto": {
        "required": [
          "assistantBlocks",
          "kind",
          "progress",
          "suggestedInputType",
          "topic"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/OnboardingTurnKind"
          },
          "assistantBlocks": { },
          "topic": {
            "$ref": "#/components/schemas/OnboardingTopic"
          },
          "suggestedInputType": {
            "$ref": "#/components/schemas/SuggestedInputType"
          },
          "progress": {
            "$ref": "#/components/schemas/OnboardingProgressDto"
          },
          "planId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PrimaryGoal": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "PrimaryGoalAnswer": {
        "required": [
          "description",
          "goal"
        ],
        "type": "object",
        "properties": {
          "goal": {
            "$ref": "#/components/schemas/PrimaryGoal"
          },
          "description": {
            "type": "string",
            "description": "Runner-supplied free-text description that informed the categorical mapping."
          }
        },
        "additionalProperties": false
      },
      "PrimaryGoalInputDto": {
        "required": [
          "goal"
        ],
        "type": "object",
        "properties": {
          "goal": {
            "$ref": "#/components/schemas/PrimaryGoal"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ReviseAnswerRequestDto": {
        "required": [
          "normalizedValue",
          "topic"
        ],
        "type": "object",
        "properties": {
          "topic": {
            "$ref": "#/components/schemas/OnboardingTopic"
          },
          "normalizedValue": { }
        },
        "additionalProperties": false
      },
...
      "SubmitStructuredAnswersRequestDto": {
        "required": [
          "currentFitness",
          "idempotencyKey",
          "injuryHistory",
          "preferences",
          "primaryGoal",
          "targetEvent",
          "weeklySchedule"
        ],
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "format": "uuid"
          },
          "primaryGoal": {
            "$ref": "#/components/schemas/PrimaryGoalInputDto"
          },
          "targetEvent": {
            "$ref": "#/components/schemas/TargetEventInputDto"
          },
          "currentFitness": {
            "$ref": "#/components/schemas/CurrentFitnessInputDto"
          },
          "weeklySchedule": {
            "$ref": "#/components/schemas/WeeklyScheduleInputDto"
          },
          "injuryHistory": {
            "$ref": "#/components/schemas/InjuryHistoryInputDto"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesInputDto"
          }
        },
        "additionalProperties": false
      },
      "SuggestedInputType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "TargetEventAnswer": {
        "required": [
          "distanceKm",
          "eventDateIso",
          "eventName",
          "targetFinishTimeIso"
        ],
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string",
            "description": "Name of the goal race or event, such as 'Berlin Marathon' or 'Local 10K'."
          },
          "distanceKm": {
            "type": "number",
            "description": "Target distance in kilometers for the event.",
            "format": "double"
          },
          "eventDateIso": {
            "type": "string",
            "description": "Target event date in ISO-8601 calendar form (yyyy-MM-dd)."
          },
          "targetFinishTimeIso": {
            "type": "string",
            "description": "Optional target finishing time as ISO-8601 duration (e.g. PT1H45M30S). Null if the runner has no specific time goal.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "description": "Session cookie issued by /api/v1/auth/login.",
        "name": "__Host-RunCoach",
        "in": "cookie"
      },
      "bearerAuth": {
        "type": "http",
        "description": "JWT bearer token (Authorization: Bearer ...). Either scheme satisfies the CookieOrBearer policy.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}